requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
better-profanity==0.7.0
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.8
ctransformers==0.2.27
dataclasses-json==0.6.6
distro==1.9.0
Django==5.0.6
django-auditlog==3.0.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-api-key==3.0.0
djangorestframework-simplejwt==5.3.1
drf-api-logger==1.1.16
ecdsa==0.19.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.5.0
greenlet==3.0.3
groq==0.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-text-splitters==0.2.0
langsmith==0.1.59
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openpyxl==3.1.5
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
psycopg2==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pytest==8.2.2
pytest-django==4.8.0
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
razorpay==1.4.2
requests==2.31.0
rsa==4.9
setuptools==65.5.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
tenacity==8.3.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
